+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 **Terraform**
 - Add `out` parameter to `plan` method.
 - Fix `json` parameter for `validate` method.
+- Add `return` parameter to `validate` method.
 
 ### 1.6.2
 **General**